refactor: rename datepicker to date-picker
