+++ packages/wiz-ui-next/src/components/base/inputs/index.ts
@@ -6,7 +6,7 @@ export * from "./radio";
 export * from "./toggle-switch";
 export * from "./panel-switch";
 export * from "./timepicker";
-export * from "./datepicker";
+export * from "./date-picker";
 export * from "./upload";
 export * from "./search";
 export * from "./range-datepicker";
